+++ image_transport/include/image_transport/subscriber_plugin.h
@@ -123,7 +123,7 @@ public:
   /**
    * \brief Returns the number of publishers this subscriber is connected to.
    */
-  virtual uint32_t getNumPublishers() const = 0;
+  virtual size_t getNumPublishers() const = 0;
 
   /**
    * \brief Unsubscribe the callback associated with this SubscriberPlugin.
